Missed tank + teeth block pushing in the last commit.
The CMM_NOPUSHING flag is no longer used, but i have left it in the code for some reason. The CMM_TELEPORTPUSH flag is no longer used for the purpose it was created for--the stacked block bug--but it is still being used for other purposes. Rename it? Maybe CMM_TELEPORING. There are 3 flags (CMM_NOLEAVECHECK, CMM_NODEFERBUTTONS, CMM_NOEXPOSEWALLS) which are only used when teleporting. Combine them?