Get tworld compiling under SDL2
Almost immediately segfaults, but it's a start.

SDL_DisplayFormatAlpha was copied from the SDL source history. It's just
here as scaffolding. It will be deleted later.

The Migration Guide on the SDL Wiki was a good resource.
https://wiki.libsdl.org/MigrationGuide.

Surfaces are always in RAM now and there's a new Renderer API that does
hardware-accelerated rendering. I'm sticking to Surfaces for now because
that's less churn.