change the input key and correctly return from displaylist
CmdQuit is what ALT-F4 is mapped to - in most cases it just results in an
exit(0) call.

CmdQuitLevel is what the 'q' key is mapped to - it usually does a clean
exit from whatever menu or screen we're currently in.

i'd prefer that we exit cleanly, so make input return CmdQuitLevel
instead of CmdQuit.